feature: 추천 티 리스트 페이지 마크업
